Rename o.m.a.i.option.ReportingPolicy to o.m.a.i.prism.ReportingPolicyPrism and add missing prism enum tests.
